# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: A²ES - AI for Energy & Sustainability API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /equipment/{id}/analysis:
    get:
      operationId: getEquipmentAnalysis
      tags: [analysis]
      summary: Get latest analysis for equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "404":
          description: No analysis found
  /analysis:
    post:
      operationId: analyzeEquipment
      tags: [analysis]
      summary: Analyze equipment with AI (text/image description)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeEquipmentBody"
      responses:
        "200":
          description: AI analysis result with equipment suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
  /analysis/{id}/confirm:
    post:
      operationId: confirmAnalysis
      tags: [analysis]
      summary: Confirm machine type selection and finalize analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmAnalysisBody"
      responses:
        "200":
          description: Finalized analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
...
components:
  schemas:
...
    AnalyzeEquipmentBody:
      type: object
      properties:
        textDescription:
          type: string
          nullable: true
        imageBase64:
          type: string
          nullable: true
        imageMimeType:
          type: string
          nullable: true
        equipmentData:
          $ref: "#/components/schemas/EquipmentInputData"
        equipmentId:
          type: integer
          nullable: true
    EquipmentInputData:
      type: object
      properties:
        type:
          type: string
          nullable: true
        brand:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        yearManufactured:
          type: integer
          nullable: true
        voltage:
          type: string
          nullable: true
        current:
          type: string
          nullable: true
        power:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        size:
          type: string
          nullable: true
        hoursPerDay:
          type: number
          nullable: true
        daysPerWeek:
          type: number
          nullable: true
        sector:
          type: string
          nullable: true
        apparentCondition:
          type: string
          nullable: true
        noiseObserved:
          type: boolean
          nullable: true
        heatingObserved:
          type: boolean
          nullable: true
        vibrationObserved:
          type: boolean
          nullable: true
        operatorNotes:
          type: string
          nullable: true
    ConfirmAnalysisBody:
      type: object
      properties:
        selectedSuggestionIndex:
          type: integer
      required:
        - selectedSuggestionIndex
    AnalysisResult:
      type: object
      properties:
        id:
          type: integer
        equipmentId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [pending, suggestions, confirmed, failed]
        machineSuggestions:
          type: array
          items:
            $ref: "#/components/schemas/MachineSuggestion"
        selectedSuggestionIndex:
          type: integer
          nullable: true
        identifiedType:
          type: string
          nullable: true
        identifiedBrand:
          type: string
          nullable: true
        identifiedModel:
          type: string
          nullable: true
        estimatedPowerKw:
          type: number
          nullable: true
        estimatedConsumptionHourlyKwh:
          type: number
          nullable: true
        estimatedConsumptionDailyKwh:
          type: number
          nullable: true
        estimatedConsumptionMonthlyKwh:
          type: number
          nullable: true
        estimatedCo2HourlyKg:
          type: number
          nullable: true
        estimatedCo2MonthlyKg:
          type: number
          nullable: true
        efficiencyLevel:
          type: string
          nullable: true
          enum: [inefficient, average, optimized]
        problemsFound:
          type: array
          items:
            type: string
        retrofitRecommendations:
          type: array
          items:
            $ref: "#/components/schemas/RetrofitRecommendation"
        estimatedRetrofitCostBrl:
          type: number
          nullable: true
        estimatedMonthlySavingsBrl:
          type: number
          nullable: true
        estimatedAnnualSavingsBrl:
          type: number
          nullable: true
        estimatedRoiMonths:
          type: number
          nullable: true
        estimatedConsumptionReductionPct:
          type: number
          nullable: true
        estimatedCo2ReductionKgMonth:
          type: number
          nullable: true
        aiSummary:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - machineSuggestions
        - problemsFound
        - retrofitRecommendations
        - createdAt
    MachineSuggestion:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        confidence:
          type: string
          enum: [high, medium, low]
      required:
        - title
        - description
        - confidence
    RetrofitRecommendation:
      type: object
      properties:
        category:
          type: string
        item:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
        estimatedCostBrl:
          type: number
          nullable: true
      required:
        - category
        - item
        - description
        - priority
...