# lib/api-zod/src/generated/types/analyzeEquipmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A²ES - AI for Energy & Sustainability API
 * OpenAPI spec version: 0.1.0
 */
import type { EquipmentInputData } from "./equipmentInputData";

export interface AnalyzeEquipmentBody {
  textDescription?: string | null;
  imageBase64?: string | null;
  imageMimeType?: string | null;
  equipmentData?: EquipmentInputData;
  equipmentId?: number | null;
}