# lib/api-zod/src/generated/types/equipmentInputData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A²ES - AI for Energy & Sustainability API
 * OpenAPI spec version: 0.1.0
 */

export interface EquipmentInputData {
  type?: string | null;
  brand?: string | null;
  model?: string | null;
  yearManufactured?: number | null;
  voltage?: string | null;
  current?: string | null;
  power?: string | null;
  color?: string | null;
  size?: string | null;
  hoursPerDay?: number | null;
  daysPerWeek?: number | null;
  sector?: string | null;
  apparentCondition?: string | null;
  noiseObserved?: boolean | null;
  heatingObserved?: boolean | null;
  vibrationObserved?: boolean | null;
  operatorNotes?: string | null;
}