# lib/api-zod/src/generated/types/machineSuggestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A²ES - AI for Energy & Sustainability API
 * OpenAPI spec version: 0.1.0
 */
import type { MachineSuggestionConfidence } from "./machineSuggestionConfidence";

export interface MachineSuggestion {
  title: string;
  description: string;
  confidence: MachineSuggestionConfidence;
}